pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-ssh"
version = "1.0.0"
description = "A lightweight, no-nonsense SSH intrusion detection system using journald and ufw."
readme = "README.md"
authors = [
  { name = "Yash Potdar", email = "yashyogeshpotdar-py" },
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: System :: Networking :: Firewalls",
]
dependencies = [
    "pyyaml",
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/yashpotdar-py/sentinel-ssh"
"Bug Tracker" = "https://github.com/yashpotdar-py/sentinel-ssh/issues"

[tool.setuptools]
packages = ["sentinel"]
